auth-api/requirements.txt
flask
jwcrypto
pynacl